Revert "moving temp and out to misc"
This reverts commit 458ae32e060acb2ef037afd157a690d8aa639f79.
